+++ tests/testlib/s2n_stuffer_hex.c
@@ -40,7 +40,7 @@ static int s2n_stuffer_read_n_bits_hex(struct s2n_stuffer *stuffer, uint8_t n, u
     /* Start with u = 0 */
     *u = 0;
 
-    for (int i = 0; i < b.size; i++) {
+    for (size_t i = 0; i < b.size; i++) {
         *u <<= 4;
         if (b.data[i] >= '0' && b.data[i] <= '9') {
             *u |= b.data[i] - '0';
@@ -60,7 +60,7 @@ int s2n_stuffer_read_hex(struct s2n_stuffer *stuffer, struct s2n_stuffer *out, u
 {
     gte_check(s2n_stuffer_space_remaining(out), n);
 
-    for (int i = 0; i < n; i++) {
+    for (size_t i = 0; i < n; i++) {
         uint8_t c;
         GUARD(s2n_stuffer_read_uint8_hex(stuffer, &c));
         GUARD(s2n_stuffer_write_uint8(out, c));
@@ -73,7 +73,7 @@ int s2n_stuffer_write_hex(struct s2n_stuffer *stuffer, struct s2n_stuffer *in, u
 {
     gte_check(s2n_stuffer_space_remaining(stuffer), n * 2);
 
-    for (int i = 0; i < n; i++) {
+    for (size_t i = 0; i < n; i++) {
         uint8_t c;
         GUARD(s2n_stuffer_read_uint8(in, &c));
         GUARD(s2n_stuffer_write_uint8_hex(stuffer, c));
@@ -130,7 +130,7 @@ static int s2n_stuffer_write_n_bits_hex(struct s2n_stuffer *stuffer, uint8_t n, 
 
     lte_check(n, 64);
 
-    for (int i = b.size; i > 0; i--) {
+    for (size_t i = b.size; i > 0; i--) {
         b.data[i - 1] = hex[u & 0x0f];
         u >>= 4;
     }
@@ -168,7 +168,7 @@ int s2n_stuffer_alloc_ro_from_hex_string(struct s2n_stuffer *stuffer, const char
 
     GUARD(s2n_stuffer_alloc(stuffer, strlen(str) / 2));
 
-    for (int i = 0; i < strlen(str); i += 2) {
+    for (size_t i = 0; i < strlen(str); i += 2) {
         uint8_t u = 0;
 
         if (str[i] >= '0' && str[i] <= '9') {
