+++ tests/unit/s2n_mem_usage_test.c
@@ -116,7 +116,7 @@ int main(int argc, char **argv)
     EXPECT_NOT_EQUAL(vm_data_initial, -1);
 
     /* Allocate all connections */
-    for (int i = 0; i < connectionsToUse; i++)
+    for (size_t i = 0; i < connectionsToUse; i++)
     {
         struct s2n_connection *client_conn;
         EXPECT_NOT_NULL(client_conn = s2n_connection_new(S2N_CLIENT));
@@ -134,7 +134,7 @@ int main(int argc, char **argv)
     ssize_t vm_data_after_allocation = get_vm_data_size();
     EXPECT_NOT_EQUAL(vm_data_after_allocation, -1);
 
-    for (int i = 0; i < connectionsToUse; i++) {
+    for (size_t i = 0; i < connectionsToUse; i++) {
         EXPECT_SUCCESS(s2n_connections_set_io_pair(clients[ i ], servers[ i ], &io_pair));
 
         EXPECT_SUCCESS(s2n_negotiate_test_server_and_client(servers[i], clients[i]));
@@ -143,21 +143,21 @@ int main(int argc, char **argv)
     ssize_t vm_data_after_handshakes = get_vm_data_size();
     EXPECT_NOT_EQUAL(vm_data_after_handshakes, -1);
 
-    for (int i = 0; i < connectionsToUse; i++) {
+    for (size_t i = 0; i < connectionsToUse; i++) {
         EXPECT_SUCCESS(s2n_connection_free_handshake(servers[i]));
         EXPECT_SUCCESS(s2n_connection_free_handshake(clients[i]));
     }
     ssize_t vm_data_after_free_handshake = get_vm_data_size();
     EXPECT_NOT_EQUAL(vm_data_after_free_handshake, -1);
 
-    for (int i = 0; i < connectionsToUse; i++) {
+    for (size_t i = 0; i < connectionsToUse; i++) {
         EXPECT_SUCCESS(s2n_connection_release_buffers(servers[i]));
         EXPECT_SUCCESS(s2n_connection_release_buffers(clients[i]));
     }
     ssize_t vm_data_after_release_buffers = get_vm_data_size();
     EXPECT_NOT_EQUAL(vm_data_after_release_buffers, -1);
 
-    for (int i = 0; i < connectionsToUse; i++) {
+    for (size_t i = 0; i < connectionsToUse; i++) {
         EXPECT_SUCCESS(s2n_connection_free(clients[i]));
         EXPECT_SUCCESS(s2n_connection_free(servers[i]));
     }
